requirements.txt
Flask
Flask-WTF
Flask-Cors
pymongo
python-dotenv
python-docx